requirements.txt
python-telegram-bot>=20.7
playwright>=1.40.0
beautifulsoup4>=4.12.0
deep-translator>=1.11.0
